+++ org.eclipse.buildship.core/src/main/java/org/eclipse/buildship/core/internal/console/ProcessStreamsProvider.java
@@ -28,13 +28,4 @@ public interface ProcessStreamsProvider {
      * @return the new instance
      */
     ProcessStreams createProcessStreams(ProcessDescription processDescription);
-
-    /**
-     * Searches the existing streams for the target description and creates a new one if none found.
-     *
-     * @param processDescription the backing process
-     * @return the new instance
-     */
-    ProcessStreams getOrCreateProcessStreams(ProcessDescription processDescription);
-
 }
