+++ org.eclipse.buildship.core/src/main/java/org/eclipse/buildship/core/internal/console/StdProcessStreamsProvider.java
@@ -59,8 +59,4 @@ public final class StdProcessStreamsProvider implements ProcessStreamsProvider {
         return this.stdStreams;
     }
 
-    @Override
-    public ProcessStreams getOrCreateProcessStreams(ProcessDescription processDescription) {
-        return this.stdStreams;
-    }
 }
