+++ org.eclipse.buildship.ui.test/src/main/groovy/org/eclipse/buildship/ui/internal/test/fixtures/TestProcessStreamProvider.groovy
@@ -30,13 +30,6 @@ abstract class TestProcessStreamProvider implements ProcessStreamsProvider {
         result
     }
 
-    @Override
-    public ProcessStreams getOrCreateProcessStreams(ProcessDescription processDescription) {
-        ProcessStreams result = new TestProcessStream()
-        processStreams += result
-        result
-    }
-
     static class TestProcessStream implements ProcessStreams {
 
         OutputStream confing = new ByteArrayOutputStream()
